Rename Cancel/Restore button to reflect actual use
