+++ src/components/ToggleApplicationButton/index.test.tsx
